requirements.txt
bcrypt==3.2.0
boto3==1.20.21
fastapi==0.68.2
httptools==0.2.0
iso8601==0.1.16
jsonschema==4.2.1
passlib==1.7.4
pydantic==1.8.2
uvicorn==0.15.0
websockets==10.1
python-jose==3.3.0
python-multipart==0.0.5